Simple fix for incorrect caching of version list
This fix addresses the incorrect caching of dynamic version listing
when different configurations have different configured cache expiry.
The fix is a bit simplistic: only version lists that have been
refreshed from a remote repository will be kept in the in-memory
cache.
A better solution would properly honour the CachePolicy for all
things held in the `InMemoryMetadataCache` and the
`InMemoryArtifactsCache`. One option is to move the in-memory caching
to be closer to the file-backed caching, perhaps as a single read
-through cache.