Update test to show that variant information can be used to determine the target "configuration"
For local projects, the variant name found in the resolution result matches the target configuration
name of the local component being selected. This commit adds a test to show that consumers can
actually use the same method to query both the "target configuration name" and "target variant name"
as they are the same concept.

This should allow https://github.com/gradle/dotcom/issues/119 to be implemented.

Signed-off-by: Cedric Champeau <cedric@gradle.com>